feat(groq-codegen): working groq plucking! groq-codegen now includes an API to pull groq from code
Breaking Change: `generateGroqTypes` has been renamed to `transformGroqToTypescript` and a combined implmentation has taken its place.

See the updated groq-codegen README for more info.
