+++ packages/groq-codegen/package.json
@@ -34,8 +34,11 @@
     "lint": "eslint --ext .ts,.js ./src"
   },
   "dependencies": {
-    "@babel/core": "^7.13.15",
-    "@babel/types": "^7.13.14",
+    "@babel/core": "^7.14.0",
+    "@babel/generator": "^7.14.0",
+    "@babel/types": "^7.14.0",
+    "@ricokahler/pool": "^1.1.2",
+    "glob": "^7.1.7",
     "groq-js": "^0.2.0"
   },
   "engines": {
