+++ packages/groq-codegen/src/index.ts
@@ -1,1 +1,3 @@
 export * from './generate-groq-types';
+export * from './pluck-groq-from-files';
+export * from './transform-groq-to-typescript';
