+++ packages/groq-codegen/tsconfig.json
@@ -9,5 +9,6 @@
     "strict": true,
     "target": "esnext"
   },
-  "include": ["./src"]
+  "include": ["./src"],
+  "exclude": ["**/__example-files__/**/*"]
 }
