+++ scripts/build
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-# order matters
+# order matters, thus this script
+# codegens have to go first
 yarn workspace @sanity-codegen/schema-codegen run build
-yarn workspace @sanity-codegen/cli run build
 yarn workspace @sanity-codegen/groq-codegen run build
+
+# then the rest of them are fine
+yarn workspace @sanity-codegen/cli run build
+# yarn workspace @sanity-codegen/client run build
+
+# note: `@sanity-codegen/types` and `sanity-codegen` don't require a build step
